Quiz part of flashcards
Adds a new collapsible panel "Quiz".  When expanded, disables the other
panels ("New" and "Cards"). A disabled collapsible panel is always
collapsed.  Quiz displays each card one at a time, in random order.
Quiz cards show the question, a button to check the answer, and
pagination buttons.  Once clicked, the text changes to the answer.
Can paginate forward and back through the cards during the quiz.
Pagination resets the answer, so it's always hidden when the card is
first rendered.

Quiz does not ask for any input, and as such does not bother checking
the answers (just like real flashcards - it's up to you to decide if you
really know the card or not).